Changes following instructions by owner and issues

+++ cuda_rasterizer/utils.cu
@@ -4,29 +4,30 @@
 
 // Equation (9) in "3D Gaussian Splatting as Markov Chain Monte Carlo"
 __global__ void compute_relocation(
-    int P, 
-    float* opacity_old, 
-    float* scale_old, 
-    int* N, 
-    float* binoms, 
-    int n_max, 
-    float* opacity_new, 
-    float* scale_new) 
+    int P,
+    float* opacity_old,
+    float* scale_old,
+    int* N,
+    float* binoms,
+    int n_max,
+    float* opacity_new,
+    float* scale_new)
 {
     int idx = threadIdx.x + blockIdx.x * blockDim.x;
     if (idx >= P) return;
-    
+
     int N_idx = N[idx];
     float denom_sum = 0.0f;
 
     // compute new opacity
     opacity_new[idx] = 1.0f - powf(1.0f - opacity_old[idx], 1.0f / N_idx);
-    
+
     // compute new scale
     for (int i = 1; i <= N_idx; ++i) {
         for (int k = 0; k <= (i-1); ++k) {
             float bin_coeff = binoms[(i-1) * n_max + k];
-            float term = (pow(-1, k) / sqrt(k + 1)) * pow(opacity_new[idx], k + 1);
+            // float term = (pow(-1, k) / sqrt(k + 1)) * pow(opacity_new[idx], k + 1);
+            float term = (pow(-1, k) / sqrt((float)(k + 1))) * pow(opacity_new[idx], k + 1);
             denom_sum += (bin_coeff * term);
         }
     }
